support for multiple users added
other user can access,update,delete their todos only

+++ controllers/todoControllers.js
@@ -1,8 +1,9 @@
 //models
 const Todos = require("../models/todos");
 
+// res => will contain currentUser object currently logged in
 const getAllTodos = (req, res, next) => {
-  Todos.find({})
+  Todos.find({ author: res.currentUser._id })
     .then((todos) => {
       res.status(200);
       res.setHeader("Content-Type", "application/json");
@@ -15,6 +16,7 @@ const getAllTodos = (req, res, next) => {
 };
 
 const createTodo = (req, res, next) => {
+  req.body.author = res.currentUser._id;
   Todos.create(req.body)
     .then((todo) => {
       res.status(200);
@@ -30,9 +32,14 @@ const createTodo = (req, res, next) => {
 const getTodoById = (req, res, next) => {
   Todos.findById(req.params.id)
     .then((todo) => {
-      res.status(200);
-      res.setHeader("Content-Type", "application/json");
-      res.json(todo);
+      if (todo.author == res.currentUser._id) {
+        res.status(200);
+        res.setHeader("Content-Type", "application/json");
+        res.json(todo);
+      } else {
+        res.status(401);
+        res.json({ message: "unauthorized operation" });
+      }
     })
     .catch((err) => {
       res.status(404);
@@ -41,17 +48,29 @@ const getTodoById = (req, res, next) => {
 };
 
 const updateTodo = (req, res, next) => {
-  Todos.findByIdAndUpdate(
-    req.params.id,
-    {
-      $set: req.body,
-    },
-    { new: true, useFindAndModify: false } //get updated result
-  )
+  Todos.findById(req.params.id)
     .then((todo) => {
-      res.status(200);
-      res.setHeader("Content-Type", "application/json");
-      res.json(todo);
+      if (todo.author == res.currentUser._id) {
+        Todos.findByIdAndUpdate(
+          { _id: req.params.id },
+          {
+            $set: req.body,
+          },
+          { new: true, useFindAndModify: false } //get updated result
+        )
+          .then((todo) => {
+            res.status(200);
+            res.setHeader("Content-Type", "application/json");
+            res.json(todo);
+          })
+          .catch((err) => {
+            res.status(404);
+            res.json({ message: "unable to update", error: err });
+          });
+      } else {
+        res.status(401);
+        res.json({ message: "unauthorized operation" });
+      }
     })
     .catch((err) => {
       res.status(404);
@@ -59,11 +78,26 @@ const updateTodo = (req, res, next) => {
     });
 };
 const deleteTodo = (req, res, next) => {
-  Todos.findByIdAndRemove(req.params.id)
-    .then((response) => {
-      res.status(200);
-      res.setHeader("Content-Type", "application/json");
-      res.json({ status: "Todo deleted successfully", response: response });
+  Todos.findById(req.params.id)
+    .then((todo) => {
+      if (todo.author == res.currentUser._id) {
+        Todos.findByIdAndRemove(req.params.id)
+          .then((response) => {
+            res.status(200);
+            res.setHeader("Content-Type", "application/json");
+            res.json({
+              status: "Todo deleted successfully",
+              response: response,
+            });
+          })
+          .catch((err) => {
+            res.status(404);
+            res.json({ message: "unable to delete", error: err });
+          });
+      } else {
+        res.status(401);
+        res.json({ message: "unauthorized operation" });
+      }
     })
     .catch((err) => {
       res.status(404);
